RavenDB-19057 - fixing PR comments#1
